remove unnecessary class
