+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/my/MyChannelAccessor.scala
@@ -1,12 +1,6 @@
 package x7c1.linen.repository.channel.my
 
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.database.mixin.MyChannelRecord
-import x7c1.linen.database.struct.{AccountIdentifiable, ChannelDeletable, ChannelIdentifiable}
-import x7c1.linen.repository.account.ClientAccount
-import x7c1.linen.repository.date.Date
-import x7c1.wheat.modern.database.selector.CursorConvertible
-import x7c1.wheat.modern.database.selector.presets.{CanTraverseEntity, DefaultProvidable}
 import x7c1.wheat.modern.sequence.Sequence
 
 trait MyChannelAccessor extends Sequence[MyChannelRow]{
@@ -14,41 +8,15 @@ trait MyChannelAccessor extends Sequence[MyChannelRow]{
 
 object MyChannelAccessor {
   def createForDebug(helper: DatabaseHelper, accountId: Long): MyChannelAccessor = {
-    ClosableMyChannelAccessor.create(helper, ClientAccount(accountId)) match {
+    helper.selectorOf[MyChannel] traverseOn accountId match {
       case Left(e) => throw e
-      case Right(accessor) => accessor
+      case Right(accessor) => apply(accessor)
     }
   }
+  private def apply(sequence: Sequence[MyChannelRow]): MyChannelAccessor =
+    new MyChannelAccessor {
+      override def findAt(position: Int) = sequence findAt position
+      override def length = sequence.length
+    }
+
 }
-
-sealed trait MyChannelRow
-
-case class MyChannel(
-  channelId: Long,
-  name: String,
-  description: String,
-  createdAt: Date,
-  isSubscribed: Boolean ) extends MyChannelRow
-
-object MyChannel {
-  implicit object id extends ChannelIdentifiable[MyChannel]{
-    override def toId = _.channelId
-  }
-  implicit object deletable extends ChannelDeletable[MyChannel](_.channelId)
-
-  implicit object convertible extends CursorConvertible[MyChannelRecord, MyChannel]{
-    override def fromCursor = cursor =>
-      MyChannel(
-        channelId = cursor._id,
-        name = cursor.name,
-        description = cursor.description,
-        createdAt = cursor.created_at.typed,
-        isSubscribed = cursor.subscribed == 1
-      )
-  }
-  implicit object traversable extends CanTraverseEntity[AccountIdentifiable, MyChannelRecord, MyChannel]
-
-  implicit object providable extends DefaultProvidable[AccountIdentifiable, MyChannel]
-}
-
-case class MyChannelFooter() extends MyChannelRow
